restaurant_backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.9.2
pydantic==2.8.0
pydantic[email]==2.8.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
aiofiles==24.1.0
